pyproject.toml
[tool.poetry]
name = "chat-with-pdf"
version = "0.1.0"
description = ""
authors = ["Makoto Tanji <tanji.makoto@gmail.com>"]
readme = "README.md"
packages = [{include = "chat_with_pdf"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.21.0"
llama-index = "^0.5.12"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
